fix auth page and refactor code
