+++ src/components/SlideBar/index.tsx
@@ -1,60 +1,1 @@
-// import * as React from 'react';
-// import Button from '@mui/material/Button';
-// import Snackbar from '@mui/material/Snackbar';
-// import Fade from '@mui/material/Fade';
-// import Slide, { SlideProps } from '@mui/material/Slide';
-// import { TransitionProps } from '@mui/material/transitions';
-
-// function SlideTransition(props: SlideProps) {
-//   return <Slide {...props} direction="up" />;
-// }
-
-
-// export default function TransitionsSnackbar() {
-//   const [state, setState] = React.useState<{
-//     open: boolean;
-//     Transition: React.ComponentType<
-//       TransitionProps & {
-//         children: React.ReactElement<any, any>;
-//       }
-//     >;
-//   }>({
-//     open: false,
-//     Transition: Fade,
-//   });
-
-//   const handleClick =
-//     (
-//       Transition: React.ComponentType<
-//         TransitionProps & {
-//           children: React.ReactElement<any, any>;
-//         }
-//       >,
-//     ) =>
-//     () => {
-//       setState({
-//         open: true,
-//         Transition,
-//       });
-//     };
-
-//   const handleClose = () => {
-//     setState({
-//       ...state,
-//       open: false,
-//     });
-//   };
-
-//   return (
-//     <div>
-//       <Button onClick={handleClick(SlideTransition)}>Slide Transition</Button>
-//       <Snackbar
-//         open={state.open}
-//         onClose={handleClose}
-//         TransitionComponent={state.Transition}
-//         message="I love snacks"
-//         key={state.Transition.name}
-//       />
-//     </div>
-//   );
-// }
+export {default as SlideForgot} from './SlideForgot/index';
